fix: rename link button to text button
