+++ src/templates/GuestMain.template.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 // components
 import AppTitleComponent from "@src/components/molecules/AppTitle.component";
 import { Button } from "@src/components/atoms/Button";
-import { TextButton } from "@src/components/atoms/LinkButton";
+import { TextButton } from "@src/components/atoms/TextButton";
 
 // styles
 import SGuestMain from "@src/styles/template/GuestMain.styles";
